requirements.txt
flask==3.0.3
supabase==2.4.6
PyJWT==2.8.0
bcrypt==4.1.3
requests==2.31.0
pydub==0.25.1